refactor: change external tour api package structure
